+++ react/country/VEN.js
@@ -222,6 +222,7 @@ export default {
       postalCodeAPI: false,
       regex: /^\d{4}$/,
       size: 'small',
+      required: true,
     },
     {
       name: 'street',
@@ -290,11 +291,14 @@ export default {
       valueIn: 'long_name',
       types: ['postal_code'],
     },
+
     complement: {
       valueIn: 'long_name',
       types: ['street_number'],
     },
+
     street: { valueIn: 'long_name', types: ['route'] },
+
     neighborhood: {
       valueIn: 'long_name',
       types: [
@@ -307,13 +311,19 @@ export default {
       ],
       required: false,
     },
+
     state: {
       valueIn: 'long_name',
       types: ['administrative_area_level_1'],
     },
+
     city: {
       valueIn: 'long_name',
       types: ['administrative_area_level_2', 'locality'],
+    },
+
+    receiverName: {
+      required: true,
     },
   },
   summary: [