+++ react/country/FRA.js
@@ -71,21 +71,29 @@ export default {
       types: ['postal_code'],
       required: false,
     },
+
     street: {
       valueIn: 'long_name',
       types: ['route'],
       handler: (address, googleAddress) => {
         address.street = { value: googleAddress.name }
+
         return address
       },
     },
+
     city: {
       valueIn: 'long_name',
       types: ['locality'],
     },
+
     state: {
       valueIn: 'short_name',
       types: ['administrative_area_level_1'],
+    },
+
+    receiverName: {
+      required: true,
     },
   },
   summary: [