+++ react/country/__mocks__/useOneLevel.js
@@ -75,7 +75,7 @@ export default {
       valueIn: 'long_name',
       types: ['postal_code'],
       required: false,
-      handler: address => {
+      handler: (address) => {
         if (
           !address.state ||
           !address.city ||
@@ -97,20 +97,36 @@ export default {
         return address
       },
     },
-    number: { valueIn: 'long_name', types: ['street_number'], required: false, notApplicable: true, handler: address => {
-      return {...address,
-        number: {...address['number'], notApplicable: true}}
-    } },
+
+    number: {
+      valueIn: 'long_name',
+      types: ['street_number'],
+      required: false,
+      notApplicable: true,
+      handler: (address) => {
+        return {
+          ...address,
+          number: { ...address.number, notApplicable: true },
+        }
+      },
+    },
+
     street: { valueIn: 'long_name', types: ['route'], required: false },
+
     state: {
       valueIn: 'long_name',
       types: ['administrative_area_level_1'],
       required: false,
     },
+
     city: {
       valueIn: 'long_name',
       types: ['administrative_area_level_2', 'locality'],
       required: false,
+    },
+
+    receiverName: {
+      required: true,
     },
   },
 }